+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.1.1
+- `export_question_bulk_filter_values` will return successful paths if there were error paths.
+- Add `export_question_bulk_filter_values` and `export_question` to the `__init__.py` so you can import shorter.
+
+# 1.1.0
+Fix custom_retry_errors is not used in async_query.
+
 # 1.0.9
 Add check_retry_errors function for checking static errors and dynamic errors.